Move FOM functions into dedicated module `fom`
